refactor: 코드리뷰 반영하기 (파일명 수정, 이상한 문법 수정)
